refactor: moving the automation validation functions into that package
