+++ azurerm/internal/services/automation/automation_connection_certificate_resource.go
@@ -51,7 +51,7 @@ func resourceArmAutomationConnectionCertificate() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateAutomationAccountName(),
+				ValidateFunc: validate.AutomationAccountName(),
 			},
 
 			"automation_certificate_name": {
