+++ azurerm/internal/services/automation/automation_connection_classic_certificate_resource.go
@@ -51,7 +51,7 @@ func resourceArmAutomationConnectionClassicCertificate() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateAutomationAccountName(),
+				ValidateFunc: validate.AutomationAccountName(),
 			},
 
 			"subscription_id": {
