+++ azurerm/internal/services/automation/automation_connection_resource.go
@@ -52,7 +52,7 @@ func resourceArmAutomationConnection() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateAutomationAccountName(),
+				ValidateFunc: validate.AutomationAccountName(),
 			},
 
 			"type": {
