+++ azurerm/internal/services/automation/automation_connection_service_principal_resource.go
@@ -51,7 +51,7 @@ func resourceArmAutomationConnectionServicePrincipal() *schema.Resource {
 				Type:         schema.TypeString,
 				Required:     true,
 				ForceNew:     true,
-				ValidateFunc: azure.ValidateAutomationAccountName(),
+				ValidateFunc: validate.AutomationAccountName(),
 			},
 
 			"application_id": {
