+++ azurerm/internal/services/automation/helper/automation_job_schedule.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/hashcode"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 	uuid "github.com/satori/go.uuid"
-	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/azure"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/automation/validate"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/utils"
 )
 
@@ -24,7 +24,7 @@ func JobScheduleSchema() *schema.Schema {
 				"schedule_name": {
 					Type:         schema.TypeString,
 					Required:     true,
-					ValidateFunc: azure.ValidateAutomationScheduleName(),
+					ValidateFunc: validate.AutomationScheduleName(),
 				},
 
 				"parameters": {
